introduce-corpus-controller(move): Hand external-corpus operations over to CorpusController
Pure physical move per MECH_COMMIT_SPLIT. Cut @staticmethod
add_external_corpus / remove_external_corpus / list_external_corpora
from TokenizerControlMixin; paste into CorpusController (drop
@staticmethod, replace ``self: "CorpusController"`` -> plain ``self``).
Caller prefix replacement in http_server.py:
``TokenizerManager.<method>(_global_state.tokenizer_manager.corpus_controller, ...)``
-> ``_global_state.tokenizer_manager.corpus_controller.<method>(...)``.

Refactor chain ID: introduce-corpus-controller-move
PR-Title: Extract CorpusController from TokenizerManager
