+++ python/sglang/srt/entrypoints/http_server.py
@@ -816,8 +816,8 @@ async def add_external_corpus(request: Request):
             {"success": False, "message": str(e)},
             status_code=HTTPStatus.BAD_REQUEST,
         )
-    result = await TokenizerManager.add_external_corpus(
-        _global_state.tokenizer_manager.corpus_controller, obj
+    result = (
+        await _global_state.tokenizer_manager.corpus_controller.add_external_corpus(obj)
     )
     return ORJSONResponse(
         {
@@ -841,8 +841,10 @@ async def remove_external_corpus(request: Request):
             {"success": False, "message": "corpus_id is required."},
             status_code=HTTPStatus.BAD_REQUEST,
         )
-    result = await TokenizerManager.remove_external_corpus(
-        _global_state.tokenizer_manager.corpus_controller, corpus_id
+    result = (
+        await _global_state.tokenizer_manager.corpus_controller.remove_external_corpus(
+            corpus_id
+        )
     )
     return ORJSONResponse(
         {"success": result.success, "message": result.message},
@@ -854,8 +856,8 @@ async def remove_external_corpus(request: Request):
 @auth_level(AuthLevel.ADMIN_OPTIONAL)
 async def list_external_corpora():
     """List all active external corpora."""
-    result = await TokenizerManager.list_external_corpora(
-        _global_state.tokenizer_manager.corpus_controller,
+    result = (
+        await _global_state.tokenizer_manager.corpus_controller.list_external_corpora()
     )
     return ORJSONResponse(
         {
