+++ python/sglang/srt/managers/tokenizer_control_mixin.py
@@ -8,7 +8,6 @@ import fastapi
 
 from sglang.srt.managers.communicator import FanOutCommunicator
 from sglang.srt.managers.io_struct import (
-    AddExternalCorpusReqInput,
     AddExternalCorpusReqOutput,
     AttachHiCacheStorageReqInput,
     AttachHiCacheStorageReqOutput,
@@ -32,14 +31,12 @@ from sglang.srt.managers.io_struct import (
     InitWeightsSendGroupForRemoteInstanceReqInput,
     InitWeightsSendGroupForRemoteInstanceReqOutput,
     InitWeightsUpdateGroupReqOutput,
-    ListExternalCorporaReqInput,
     ListExternalCorporaReqOutput,
     LoRAUpdateOutput,
     ProfileReq,
     ProfileReqOutput,
     ProfileReqType,
     ReleaseMemoryOccupationReqOutput,
-    RemoveExternalCorpusReqInput,
     RemoveExternalCorpusReqOutput,
     ResumeMemoryOccupationReqOutput,
     SendWeightsToRemoteInstanceReqInput,
@@ -113,117 +110,6 @@ class TokenizerControlMixin:
             setattr(self, f"{name}_communicator", comm)
             dispatch_pairs.append((resp_type, comm.handle_recv))
         self._result_dispatcher += TypeBasedDispatcher(dispatch_pairs)
-
-    @staticmethod
-    async def add_external_corpus(
-        self: "CorpusController", obj: AddExternalCorpusReqInput
-    ) -> AddExternalCorpusReqOutput:
-        self.auto_create_handle_loop()
-        if self.config.speculative_algorithm != "NGRAM":
-            return AddExternalCorpusReqOutput(
-                success=False,
-                message="Ngram speculative decoding is not enabled.",
-            )
-        truncated = False
-        try:
-            if not obj.corpus_id:
-                import uuid
-
-                obj.corpus_id = uuid.uuid4().hex
-            if obj.file_path is not None:
-                from sglang.srt.speculative.cpp_ngram.external_corpus import (
-                    iter_external_corpus_chunks,
-                )
-
-                max_tokens = self.config.max_external_corpus_tokens
-                obj.token_chunks = list(
-                    iter_external_corpus_chunks(
-                        obj.file_path, self.tokenizer, max_tokens
-                    )
-                )
-            elif obj.documents is not None:
-                from sglang.srt.speculative.cpp_ngram.external_corpus import (
-                    SEPARATOR_TOKEN,
-                )
-
-                max_tokens = self.config.max_external_corpus_tokens
-                token_chunks = []
-                total_tokens = 0
-                has_prev = False
-                for doc in obj.documents:
-                    if not doc:
-                        continue
-                    token_ids = list(
-                        self.tokenizer.encode(doc, add_special_tokens=False)
-                    )
-                    if not token_ids:
-                        continue
-                    if has_prev:
-                        token_ids = [SEPARATOR_TOKEN] + token_ids
-                    if total_tokens + len(token_ids) > max_tokens:
-                        truncated = True
-                        break
-                    token_chunks.append(token_ids)
-                    total_tokens += len(token_ids)
-                    has_prev = True
-                obj.token_chunks = token_chunks
-            else:
-                return AddExternalCorpusReqOutput(
-                    success=False,
-                    message="Either file_path or documents must be provided.",
-                )
-            obj.file_path = None
-            obj.documents = None
-            results = await self.add_external_corpus_communicator(obj)
-            all_success, all_message = FanOutCommunicator.merge_results(results)
-            if truncated and all_success:
-                all_message += f" (truncated: exceeded {max_tokens} token limit)"
-            return AddExternalCorpusReqOutput(
-                success=all_success,
-                corpus_id=results[0].corpus_id if all_success else "",
-                message=all_message,
-                loaded_token_count=results[0].loaded_token_count if all_success else 0,
-            )
-        except Exception as e:
-            return AddExternalCorpusReqOutput(success=False, message=str(e))
-
-    @staticmethod
-    async def remove_external_corpus(
-        self: "CorpusController", corpus_id: str
-    ) -> RemoveExternalCorpusReqOutput:
-        self.auto_create_handle_loop()
-        if self.config.speculative_algorithm != "NGRAM":
-            return RemoveExternalCorpusReqOutput(
-                success=False,
-                message="Ngram speculative decoding is not enabled.",
-            )
-        results = await self.remove_external_corpus_communicator(
-            RemoveExternalCorpusReqInput(corpus_id=corpus_id)
-        )
-        all_success, all_message = FanOutCommunicator.merge_results(results)
-        return RemoveExternalCorpusReqOutput(success=all_success, message=all_message)
-
-    @staticmethod
-    async def list_external_corpora(
-        self: "CorpusController",
-    ) -> ListExternalCorporaReqOutput:
-        self.auto_create_handle_loop()
-        if self.config.speculative_algorithm != "NGRAM":
-            return ListExternalCorporaReqOutput(
-                success=False,
-                message="Ngram speculative decoding is not enabled.",
-            )
-        results = await self.list_external_corpora_communicator(
-            ListExternalCorporaReqInput()
-        )
-        all_success, all_message = FanOutCommunicator.merge_results(results)
-        # Merge corpus token counts from all DP ranks (each rank loads the same set).
-        corpus_token_counts = results[0].corpus_token_counts if all_success else {}
-        return ListExternalCorporaReqOutput(
-            success=all_success,
-            corpus_token_counts=corpus_token_counts,
-            message=all_message,
-        )
 
     async def flush_cache(
         self: TokenizerManager, timeout_s: Optional[float] = None
